+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix bench tests missing test_file var
 - Fix compatibility with older python versions for clock::now
 - Eliminated redundant `declare -F | awk` calls that were happening for every test/bench file
+- Replaced tail and process with Bash parameter expansion
 
 ## [0.25.0](https://github.com/TypedDevs/bashunit/compare/0.23.0...0.24.0) - 2025-10-05